Add RFC-based character-encoding to URL fixing
This adds a more advanced approach to fixing characters in query
parameters. It uses a parsec to analyze the structure of the string
in the param. Each character or string subset is therefore either
an reserved, unreserved, hex-encoded or invalid character.
The characters that are invalid are then encoded and the new
value is set.